Break editor into multiple files
